Refactor: move ReplicationState to replcation_state.rs
